requirements.txt
pydantic>=2.5.0
pydantic-ai>=0.1.0
langgraph>=0.2.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
